+++ include/stdint.h
@@ -11,8 +11,8 @@ typedef unsigned short      uint16_t;
 typedef unsigned int        uint32_t;
 typedef unsigned long long  uint64_t;
 
-typedef int32_t  intptr_t;
-typedef uint32_t uintptr_t;
+typedef int64_t  intptr_t;
+typedef uint64_t uintptr_t;
 
 #define NULL 0
